Extract ModernEventMatcher from EventResolverBase
